+++ app/src/androidTest/java/com/example/swiftcheckin/EventViewAdapterTest.java
@@ -11,6 +11,8 @@ import android.widget.TextView;
 import androidx.test.core.app.ApplicationProvider;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 
+import com.example.swiftcheckin.organizer.Event;
+
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
