+++ app/src/main/AndroidManifest.xml
@@ -24,7 +24,7 @@
         android:theme="@style/Theme.SwiftCheckIn"
         tools:targetApi="31">
         <activity
-            android:name=".ViewAttendeesActivity"
+            android:name=".organizer.ViewAttendeesActivity"
             android:exported="false" />
 
         <activity
@@ -40,7 +40,7 @@
             android:name=".AnnoucementActivity"
             android:exported="false" />
         <activity
-            android:name=".OrganizerActivity"
+            android:name=".organizer.OrganizerActivity"
             android:exported="false" />
         <activity
             android:name=".SettingsActivity"
@@ -49,7 +49,7 @@
             android:name=".ProfileActivity"
             android:exported="false" />
         <activity
-            android:name=".AddEventActivity"
+            android:name=".organizer.AddEventActivity"
             android:exported="true" />
         <activity
             android:name=".MainActivity"
