+++ app/src/main/java/com/example/swiftcheckin/AdminEventArrayAdapter.java
@@ -5,13 +5,12 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
-import android.widget.ImageView;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
-import com.bumptech.glide.Glide;
+import com.example.swiftcheckin.organizer.Event;
 
 import java.util.ArrayList;
 /**
