+++ app/src/main/java/com/example/swiftcheckin/EventViewAdapter.java
@@ -5,11 +5,10 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
-import android.widget.BaseAdapter;
 import android.widget.ImageView;
 import android.widget.TextView;
 
-import androidx.appcompat.app.AppCompatActivity;
+import com.example.swiftcheckin.organizer.Event;
 
 import java.util.List;
