+++ app/src/main/java/com/example/swiftcheckin/organizer/EventArrayAdapter.java
@@ -1,4 +1,4 @@
-package com.example.swiftcheckin;
+package com.example.swiftcheckin.organizer;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -12,6 +12,8 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
 import com.bumptech.glide.Glide;
+import com.example.swiftcheckin.R;
+import com.example.swiftcheckin.organizer.Event;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
