+++ app/src/main/java/com/example/swiftcheckin/organizer/EventSignUp.java
@@ -1,18 +1,10 @@
-package com.example.swiftcheckin;
-
-import static androidx.constraintlayout.helper.widget.MotionEffect.TAG;
-
-import android.content.Context;
-import android.util.Log;
-import android.widget.Toast;
+package com.example.swiftcheckin.organizer;
 
 import androidx.annotation.NonNull;
 
+import com.example.swiftcheckin.organizer.Event;
 import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.OnFailureListener;
-import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
-import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FirebaseFirestore;
